Fixed Pylint issues
I added a module docstring at the beginning of the file to provide a brief description of its purpose.
I added a module docstring at the beginning of the file to provide a brief description of its purpose.

I removed the wildcard import (from .models import *) and explicitly imported only the necessary class (CustomUser). Wildcard imports are generally discouraged because they can lead to namespace pollution and make it unclear which names are actually being used. In this case, it seems that other imports were not needed, so I removed them.
I added a docstring for the AccountsConfig class to describe its role in providing the configuration for the 'accounts' app.
I added a module docstring at the beginning of the file to provide a brief description of its purpose.

I added docstrings for the CustomUserManager, CustomUser, and Broker classes, as well as for the methods within the CustomUserManager class.

I addressed the linting warning regarding the __str__ method in the Broker class by returning a string.

I fixed the linting error in the Broker class related to the email attribute by updating the __str__ method to use self.user